\name{NPMLEDIC-package}
\alias{NPMLEDIC-package}
\alias{NPMLEDIC}
\docType{package}
\title{
\packageTitle{NPMLEDIC}
}
\description{
\packageDescription{NPMLEDIC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NPMLEDIC}
\packageIndices{NPMLEDIC}

}
\author{
\packageAuthor{NPMLEDIC}

Maintainer: \packageMaintainer{NPMLEDIC}
}
\references{
Ko ABE
}
\keyword{ package }
\examples{
#see
}
