\name{reccdf}
\alias{reccdf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get random ccdf from using Dirichlet distribution
}
\description{
get random ccdf using diriclet random number
}
\usage{
reccdf(n, alpha)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
sample size
}
  \item{alpha}{
parameters of dirichlet distribution}
}
