\name{rmlast}
\alias{rmlast}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
remove the last element of the vector
}
\description{
remove the last element of the vector
}
\usage{
rmlast(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  vector
}
}
\details{
remove the last element of the vector
}
